Allow fractional MLP multiplier
Read MLP multiplier from environment as a float so non-integer multipliers are supported, and cast the computed hidden size to int in the MLP constructor (hidden = int(mlp_mult * dim)) to ensure an integer layer size.

+++ train_gpt.py
@@ -65,7 +65,7 @@ class Hyperparameters:
     num_kv_heads = int(os.environ.get("NUM_KV_HEADS", 4))
     model_dim = int(os.environ.get("MODEL_DIM", 512))
     num_heads = int(os.environ.get("NUM_HEADS", 8))
-    mlp_mult = int(os.environ.get("MLP_MULT", 2))
+    mlp_mult = float(os.environ.get("MLP_MULT", 2))
     tie_embeddings = bool(int(os.environ.get("TIE_EMBEDDINGS", "1")))
     rope_base = float(os.environ.get("ROPE_BASE", 10000.0))
     logit_softcap = float(os.environ.get("LOGIT_SOFTCAP", 30.0))
@@ -606,7 +606,7 @@ class CausalSelfAttention(nn.Module):
 class MLP(nn.Module):
     def __init__(self, dim: int, mlp_mult: int):
         super().__init__()
-        hidden = mlp_mult * dim
+        hidden = int(mlp_mult * dim)
         
         self.fc1 = CastedLinear(dim, hidden, bias=False)
         self.fc2 = CastedLinear(dim, hidden, bias=False)  # gate
